Rename to Vulnerability10 for @jrojassonatype
