+++ src/main/java/org/cyclonedx/model/vulnerability/Rating.java
@@ -18,9 +18,9 @@
  */
 package org.cyclonedx.model.vulnerability;
 
-import org.cyclonedx.model.vulnerability.Vulnerability1_0.Score;
-import org.cyclonedx.model.vulnerability.Vulnerability1_0.ScoreSource;
-import org.cyclonedx.model.vulnerability.Vulnerability1_0.Severity;
+import org.cyclonedx.model.vulnerability.Vulnerability10.Score;
+import org.cyclonedx.model.vulnerability.Vulnerability10.ScoreSource;
+import org.cyclonedx.model.vulnerability.Vulnerability10.Severity;
 
 public class Rating
 {
