+++ src/main/java/org/cyclonedx/model/vulnerability/Vulnerability10.java
@@ -29,18 +29,18 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import org.cyclonedx.model.ExtensibleType;
 
 @JsonRootName("vulnerability")
-public class Vulnerability1_0
+public class Vulnerability10
     extends ExtensibleType implements Serializable
 {
   public static final String NAMESPACE_URI = "http://cyclonedx.org/schema/ext/vulnerability/1.0";
   public static final String PREFIX = "v";
   public static final String NAME = "vulnerability";
 
-  public Vulnerability1_0() {
+  public Vulnerability10() {
     super(PREFIX, NAME);
   }
 
-  public Vulnerability1_0(final String namespace, final String name) {
+  public Vulnerability10(final String namespace, final String name) {
     super(namespace, name);
   }
